refactor: change stub targets installation
related to #14
