+++ src/glfwcxx/window/CMakeLists.txt
@@ -1,4 +1,5 @@
 set(TARGET_NAME window)
+set(WINDOW_HELPER_API ${CMAKE_CURRENT_LIST_DIR}/include/helper)
 
 add_glfwcxx_target(
     GLFWCXX_TARGET_NAME
@@ -6,9 +7,12 @@ add_glfwcxx_target(
     TARGET_SOURCES
     Window.cpp
     WindowHints.cpp
+    WindowHelper.cpp
     PRIVATE_DEPENDENCY_TARGETS
     glfw::glfw
-    glfwcxx::common)
+    glfwcxx::common
+    PRIVATE_INCLUDE_DIRECTORIES
+    ${WINDOW_HELPER_API})
 add_glfwcxx_stub_target(
     GLFWCXX_TARGET_NAME
     ${TARGET_NAME}
@@ -16,7 +20,10 @@ add_glfwcxx_stub_target(
     WindowStub.cpp
     PRIVATE_DEPENDENCY_TARGETS
     glfwstub
-    glfwcxx::common-stub)
+    glfwcxx::common-stub
+    PUBLIC_INCLUDE_DIRECTORIES
+    ${WINDOW_HELPER_API})
 
 install_glfwcxx_target(GLFWCXX_TARGET_NAME ${TARGET_NAME})
-install_glfwcxx_stub_target(GLFWCXX_TARGET_NAME ${TARGET_NAME})
+install_glfwcxx_stub_target(GLFWCXX_TARGET_NAME ${TARGET_NAME} EXTRA_INSTALL_DIRECTORIES
+                            ${WINDOW_HELPER_API}/glfwcxx)
